Create README.md
this is my game that i developed it in android studio. 
it is about a ball and a paddle and some of block that you should to destry them by the ball.
watch out the ball do not pass the paddle, if the ball fall down then it will be in a jail!
if you win the game and destroy all of blocks then the ball will be grown and will fill the screen game!
other note! of you destroy any blocks then your ball size will be smaller!
do not worry about smalling your ball! just continue to try and after you will, there is no jail 
and no small ball in bad situation and your ball will be grown to fill every where.
you should to try in your life as this ball! do not worry and try! every thing will be changes in a incredible moments!